pyproject.toml
[project]
name = "pyramidpy-tools"
version = "0.1.0"
description = "PyramidPy - AI Agent Framework - Toolkits"
requires-python = ">=3.10.10" # twitter-api-client this could cause issues, fork and fix

dependencies = [
    "pydantic-settings>=2.5.2",
    "curl-cffi>=0.7.3",
    "rich>=13.9.2",
    "bs4>=0.0.2",
    "supabase>=2.9.0",
    "ucache>=0.1.4",
    "pgvector>=0.3.5",
    "psycopg>=3.2.3",
    "scrapyscript>=1.1.5",
    "tenacity>=8.5.0",
    "asyncpg>=0.29.0",
    "sqlalchemy>=2.0.35",
    "apify-client>=1.8.1",
    "asgiref>=3.8.1",
    "sqlmodel>=0.0.22",
    "langchain-community>=0.3.7",
    "marvin>=2.3.8",
    "chromadb>=0.5.18",
    "sqlite-s3-query>=0.0.82",
    "libsql-experimental>=0.0.41",
    "prefect==3.0.0",
    "slack-sdk>=3.33.3",
    "nanoid>=2.0.0",
    "sqlalchemy-libsql>=0.1.0",
    "sqlite-vec>=0.1.5",
    "apprise>=1.9.0",
    "redis>=5.2.1",
    "fakeredis>=2.26.2",
    "sse-starlette>=2.2.1",
    "autoevals>=0.0.111",
    "alembic>=1.14.0",
    "alembic-utils>=0.8.5",
    "sqlite-vss>=0.1.2",
    "pandas>=2.2.3",
    "python-ulid[pydantic]>=3.0.0",
    "arq>=0.26.1",
    "tweepy>=4.14.0",
    "twitter-api-client>=0.10.22",
    "pyjwt>=2.9.0",
    "vecs>=0.4.5",
    "tavily-python>=0.5.0",
    "celery>=5.2.7",
    "aiogram>=3.17.0",
    "discord-py[voice]>=2.4.0",
    "pynacl>=1.5.0",
    "itsdangerous>=2.2.0",
    "aiohttp>=3.11.2",
    "duckdb>=1.1.3",
    "opentelemetry-instrumentation-langchain>=0.35.0",
    "sqlalchemy-utils>=0.41.2",
    "dramatiq[redis,watch]>=1.17.1",
    "celery-redbeat>=2.2.0",
    "resend>=2.5.1",
    "pytest-env>=1.1.5",
    "pyramidpy_tools",
    "pillow>=11.1.0",
    "pydantic>=2.1.0",
    "controlflow",
]

[tool.uv.sources]
controlflow = {git = "https://github.com/pyramidpy/ControlFlow.git", tag = "v0.12.0"}


[project.optional-dependencies]
test = [
    "pytest-asyncio>=0.18.2,!=0.22.0,<0.23.0",
    "pytest-env>=0.8,<2.0",
    "pytest-rerunfailures>=10,<14",
    "pytest-sugar>=0.9,<2.0",
    "pytest-timeout",
    "pytest-xdist",
]

dev = [

]

[project.scripts]
pyramidpy-tools = "pyramidpy_tools:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
